Preparation for correct translation of mouse position

+++ nodelab/app/src/Luna/Studio/React/Model/Searcher.hs
@@ -1,15 +1,15 @@
 module Luna.Studio.React.Model.Searcher where
 
 import           Empire.API.Data.Node           (NodeId)
-import           Text.ScopeSearcher.QueryResult (QueryResult)
-import           Luna.Studio.Prelude
 import           Luna.Studio.Data.Vector
+import           Luna.Studio.Prelude
+import           Text.ScopeSearcher.QueryResult (QueryResult)
 
 
 
 data Searcher = Searcher
       { _visible     :: Bool
-      , _position    :: Vector2 Int
+      , _position    :: Vector2 Double
       , _selected    :: Int
       , _input       :: Text
       , _results     :: [QueryResult]
